+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [3.3.0] - 2020-02-03
 ### Added
-- A list of the contents of a given folder gan be retrieved.
+- Added get_folder_contents which returns list of the folders/files in the shared folder or one of its sub-folders.
 - Mocked SharedFile object now contains also the type of the file (file or directory).
 
 ## [3.2.0] - 2020-01-22